Move default equipment slots to a genEquipmentSlotsMap function.
