+++ d2game/d2player/inventory_grid.go
@@ -19,12 +19,6 @@ type InventoryItem interface {
 	SetInventoryGridSlot(x int, y int)
 }
 
-type EquipmentSlot struct {
-	Item InventoryItem
-	x    int
-	y    int
-}
-
 var ErrorInventoryFull = errors.New("inventory full")
 
 // Reusable grid for use with player and merchant inventory.
@@ -42,64 +36,13 @@ type ItemGrid struct {
 
 func NewItemGrid(width int, height int, originX int, originY int) *ItemGrid {
 	return &ItemGrid{
-		width:    width,
-		height:   height,
-		originX:  originX,
-		originY:  originY,
-		slotSize: 29,
-		sprites:  make(map[string]*d2ui.Sprite),
-		equipmentSlots: map[string]EquipmentSlot{
-			"LeftHand": {
-				Item: nil,
-				x:    438,
-				y:    200,
-			},
-			"RightHand": {
-				Item: nil,
-				x:    665,
-				y:    200,
-			},
-			"Helmet": {
-				Item: nil,
-				x:    549,
-				y:    133,
-			},
-			"Amulet": {
-				Item: nil,
-				x:    613,
-				y:    103,
-			},
-			"Chest": {
-				Item: nil,
-				x:    546,
-				y:    205,
-			},
-			"Belt": {
-				Item: nil,
-				x:    548,
-				y:    250,
-			},
-			"LeftRing": {
-				Item: nil,
-				x:    501,
-				y:    250,
-			},
-			"RightRing": {
-				Item: nil,
-				x:    613,
-				y:    250,
-			},
-			"Gloves": {
-				Item: nil,
-				x:    429,
-				y:    252,
-			},
-			"Shoes": {
-				Item: nil,
-				x:    661,
-				y:    252,
-			},
-		},
+		width:          width,
+		height:         height,
+		originX:        originX,
+		originY:        originY,
+		slotSize:       29,
+		sprites:        make(map[string]*d2ui.Sprite),
+		equipmentSlots: genEquipmentSlotsMap(),
 	}
 }
